refactor: Convert course stage solution model to typescript
